move VectorImportService to separate folder
