+++ test/modules/map/components/olMap/services/VectorImportService.test.js
@@ -1,7 +1,7 @@
-import { $injector } from '../../../../../src/injection';
-import { VectorGeoResource, VectorSourceType } from '../../../../../src/services/domain/geoResources';
-import { load } from '../../../../../src/modules/map/components/olMap/utils/feature.provider';
-import { iconUrlFunction, mapVectorSourceTypeToFormat, VectorImportService } from '../../../../../src/modules/map/components/olMap/VectorImportService';
+import { $injector } from '../../../../../../src/injection';
+import { VectorGeoResource, VectorSourceType } from '../../../../../../src/services/domain/geoResources';
+import { load } from '../../../../../../src/modules/map/components/olMap/utils/feature.provider';
+import { iconUrlFunction, mapVectorSourceTypeToFormat, VectorImportService } from '../../../../../../src/modules/map/components/olMap/services/VectorImportService';
 import VectorSource from 'ol/source/Vector';
